[9/n] Unify Tensor engine's future with ActorFuture
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/302

This replaces the "get_next_message()" polling that tensor engine client does with ActorFutures and Ports.

The major improvement is that the futures between actors and tensor engine are now the same, meaning they can be used together to wait for different actions.

This re-introduces a true controller actor that is reponsible for ensuring results are finished and that some things do not have an error before sending results to the ports.

The client is now no longer responsible for polling for anything, the workers and controller will make progress on their own.  The controller actor will directly send result values to Port given to it when it is ready.

`fetch_shard` family is now implemented by creating a Port and returning port.recv() as the future.

To unify further, the next PR will implement all the adverbs for free-standing remote functions by implementing `monarch.actor_mesh.send()` for free functions, and re-implement fetch_shard family of functions in terms of those primitives. This is possible because the controller can now route multiple results for the same seq to the result port.
ghstack-source-id: 294136061

Differential Revision: [D76918699](https://our.internmc.facebook.com/intern/diff/D76918699/)